==> notebooks/05_transform_raw_data_into_features_and_targets.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os
os.path.abspath(os.path.join(os.getcwd(), ".."))

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

# Now you can import from src
from src.data_utils import load_and_process_bike_data

rides = load_and_process_bike_data(year=2024)

rides

rides.info()

rides_cp = rides.copy()

rides['pickup_location_id'].unique()

rides

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides,6)
ts_data.head()

ts_data.tail()

ts_data

ts_data['pickup_location_id'].unique()

ts_data.shape

ts_data.shape

from src.data_utils import transform_ts_data_info_features_and_target

features, targets = transform_ts_data_info_features_and_target(ts_data, window_size=24*28, step_size=12)



features



targets

tabular_data = features
tabular_data["target"] = targets

from src.config import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet", engine="pyarrow")
