==> notebooks/09_feature_Reduction.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all" 

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))


import pandas as pd
from src.config import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from datetime import datetime

from src.data_utils import split_time_series_data

X_train, y_train, X_test, y_test = split_time_series_data(
    df,
    cutoff_date=datetime(2024, 9, 1, 0, 0, 0),
    target_column="target"
)

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

import numpy as np
import pandas as pd
from sklearn.base import BaseEstimator, RegressorMixin
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestRegressor
from sklearn.decomposition import PCA

class PCA_for_trides(BaseEstimator, RegressorMixin):
    def __init__(self, use_pca=False, n_components=10, random_state=42):
        """
        Args:
            use_pca: whether to apply PCA or just use raw features
            n_components: number of PCA components (only used if use_pca=True)
        """
        self.use_pca = use_pca
        self.n_components = n_components
        self.random_state = random_state
        
        # Internal objects
        self.scaler = None
        self.pca = None
        self.model = None
        self.feature_cols = None

    def fit(self, X, y):
       
        self.feature_cols = [col for col in X.columns if col.startswith("rides_t-")]
        X_filtered = X[self.feature_cols]
        
        # Scale
        self.scaler = StandardScaler()
        X_scaled = self.scaler.fit_transform(X_filtered)

        # Optional PCA
        if self.use_pca:
            self.pca = PCA(n_components=self.n_components, random_state=self.random_state)
            X_transformed = self.pca.fit_transform(X_scaled)
        else:
            X_transformed = X_scaled

        # Train model
        self.model = RandomForestRegressor(random_state=self.random_state)
        self.model.fit(X_transformed, y)
        return self

    def predict(self, X):
        # Expect full DataFrame
        X_filtered = X[self.feature_cols]
        X_scaled = self.scaler.transform(X_filtered)

        if self.use_pca:
            X_transformed = self.pca.transform(X_scaled)
        else:
            X_transformed = X_scaled

        return self.model.predict(X_transformed)


model = PCA_for_trides(use_pca = False)
model.fit(X_train,y_train)

y_pred = model.predict(X_test)
from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, y_pred)
print(f"Test MAE: {test_mae:.2f}")


model = PCA_for_trides(use_pca=True)
model.fit(X_train,y_train)


y_pred = model.predict(X_test)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, y_pred)
print(f"Test MAE: {test_mae:.2f}")

from src.experiment_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
sys.path
load_dotenv() 
log_model_to_mlflow(model, X_test, experiment_name="RandomRegressor_feature_reduced_without_pca", metric_name="mean_absolute_error", score=test_mae)
