==> notebooks/19_retraining_model.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

from src.inference import get_feature_store, fetch_predictions
import pandas as pd
from datetime import datetime, timedelta, timezone


def fetch_days_data(days):
    current_date = pd.to_datetime(datetime.now(timezone.utc))
    fetch_data_from = current_date - timedelta(days=(365+days)) 
    fetch_data_to = current_date - timedelta(days=365)
    print(fetch_data_from, fetch_data_to)
    fs = get_feature_store()
    fg = fs.get_feature_group(
        name=config.FEATURE_GROUP_NAME,
        version=1
    )

    query = fg.select_all()
    # query = query.filter((fg.pickup_hour >= fetch_data_from))
    df = query.read()
    cond = (df["pickup_hour"] >= fetch_data_from) & (df["pickup_hour"] <= fetch_data_to)
    return df[cond]

ts_data = fetch_days_data(180)

from src.data_utils import transform_ts_data_info_features_and_target

features, targets = transform_ts_data_info_features_and_target(ts_data, window_size=24*28, step_size=23)

from src.pipeline_utils import get_pipeline
pipeline = get_pipeline()
pipeline.fit(features, targets)

from sklearn.metrics import mean_absolute_error
predictions = pipeline.predict(features)

test_mae = mean_absolute_error(targets, predictions)
print(f"{test_mae:.4f}")

from src.inference import load_metrics_from_registry 

metric = load_metrics_from_registry()

metric

from src.inference import load_model_from_registry
model = load_model_from_registry()

import joblib  
import src.config
# Save the pipeline  
joblib.dump(pipeline, config.MODELS_DIR / "lgb_model.pkl")

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(features)
output_schema = Schema(targets)
model_schema = ModelSchema(input_schema=input_schema, output_schema=output_schema)

from src.inference import get_hopsworks_project
project = get_hopsworks_project()
model_registry = project.get_model_registry()
model_registry

modelv2 = model_registry.sklearn.create_model(
    name="citibike_demand_predict",
    metrics={"test_mae": test_mae},
    description="LightGBM regressor V2",
    input_example=features.sample(),
    model_schema=model_schema,
)

modelv2.save('/Users/ajit/Desktop/Citibike_prediction_aml/models')

from src.inference import load_model_from_registry
model = load_model_from_registry(-1)

models = model_registry.get_models(name=config.MODEL_NAME)

models

max(models, key=lambda model: model.version)


load_metrics_from_registry()
